Updates to the publish process
Since we've updated to much-newer versions of sbt-sonatype and sbt-pgp, a lot of the settings are now wrong. Things are now updated based on the docs of those plugins. This has allowed me to publish version 0.9.

Updated the README to reflect the new version.

Renamed plugins.sbt to match current standards. (You can tell how ancient this library is from the fact that it wasn't already named that.)